refactor: Majorly refactored the 'state' crate
In accordance with #4, this checks the `Add error handling at the base level (state crate)` and `Clean up the state crate` objectives. Most of this was accomplished by introducing the `anyhow` crate, which allows a ton of extra verbosity in error handling. I also used a variety of de-nesting techniques to conform to NASA's recommendation for max indent levels. Also added a ton of documentation.

Removed the old `run_inference.py` file and instead added a safely `.gitignore` to ensure that it never adds any client data, but still tracks the empty folder. Also flattened the `queue` directory down a level.

#4